refactor: replace viem provider/signer wrappers with direct implementations
