+++ packages/shared/src/viem-utils.ts
@@ -1,9 +1,6 @@
 /**
  * Shared viem utilities for TACo packages
  *
- * This module provides consistent viem integration utilities across all TACo packages,
- * eliminating code duplication while maintaining clean architecture.
- *
  * Features:
  * - Optional viem dependency handling with helpful error messages
  * - Dynamic import pattern that's webpack-compatible
@@ -13,15 +10,12 @@
  *
  * Usage:
  * ```typescript
- * import { checkViemAvailability, type PublicClient, ViemProviderBase } from '@nucypher/shared';
+ * import { checkViemAvailability, type PublicClient } from '@nucypher/shared';
  *
- * class MyViemProvider extends ViemProviderBase {
- *   // Add package-specific methods
- * }
+ * // Use viem clients directly with TACo adapters
+ * const tacoProvider = await createTacoProvider(viemPublicClient);
  * ```
  */
-
-import { ethers } from 'ethers';
 
 // Type helper: Use the real type from 'viem' if available, otherwise fallback to 'any'
 // This pattern preserves type safety for consumers who have 'viem' installed, but does not break for others.
@@ -116,278 +110,4 @@ export async function checkViemAvailability(): Promise<void> {
       'viem is required for viem wrapper functions. Install it with: npm install viem@^2.0.0',
     );
   }
-}
-
-/**
- * Base provider class that implements common viem-to-ethers provider wrapper functionality
- *
- * This class contains all the shared implementation logic between taco and taco-auth packages.
- * Package-specific provider classes can extend this to add additional methods.
- */
-export abstract class ViemProviderBase {
-  protected viemPublicClient: PublicClient;
-
-  readonly _network: Promise<ethers.providers.Network>;
-  readonly formatter?: undefined = undefined;
-
-  constructor(viemPublicClient: PublicClient) {
-    this.viemPublicClient = viemPublicClient;
-    // Initialize network for ethers compatibility
-    this._network = this.getNetwork();
-  }
-
-  async getNetwork(): Promise<ethers.providers.Network> {
-    const chainId = await this.viemPublicClient.getChainId();
-    const name = this.viemPublicClient.chain?.name || `chain-${chainId}`;
-    return {
-      name,
-      chainId,
-    };
-  }
-
-  async getBlockNumber(): Promise<number> {
-    return Number(await this.viemPublicClient.getBlockNumber());
-  }
-
-  async getBalance(
-    address: string,
-    blockTag?: string | number,
-  ): Promise<ethers.BigNumber> {
-    const address_0x = address as `0x${string}`;
-    let balance: bigint;
-
-    if (
-      blockTag === 'latest' ||
-      blockTag === 'pending' ||
-      blockTag === 'earliest' ||
-      blockTag === 'safe' ||
-      blockTag === 'finalized'
-    ) {
-      // Use blockTag for predefined string values
-      balance = await this.viemPublicClient.getBalance({
-        address: address_0x,
-        blockTag: blockTag,
-      });
-    } else if (
-      typeof blockTag === 'number' ||
-      (typeof blockTag === 'string' && blockTag.startsWith('0x'))
-    ) {
-      // Use blockNumber for hex string values
-      balance = await this.viemPublicClient.getBalance({
-        address: address_0x,
-        blockNumber: BigInt(blockTag),
-      });
-    } else {
-      // Default case (undefined or unrecognized) - use latest
-      balance = await this.viemPublicClient.getBalance({
-        address: address_0x,
-      });
-    }
-    return ethers.BigNumber.from(balance.toString());
-  }
-
-  /**
-   * @remarks The _blockTag?: string | number parameter is not used in TACo, so it is not implemented
-   */
-  async getTransactionCount(address: string): Promise<number> {
-    return await this.viemPublicClient.getTransactionCount({
-      address: address as `0x${string}`,
-    });
-  }
-
-  /**
-   * @remarks The _blockTag?: string | number parameter is not used in TACo, so it is not implemented
-   */
-  async getCode(address: string): Promise<`0x${string}` | undefined> {
-    return await this.viemPublicClient.getCode({
-      address: address as `0x${string}`,
-    });
-  }
-
-  /**
-   * @remarks The _blockTag?: string | number parameter is not used in TACo, so it is not implemented
-   */
-  async call(
-    transaction: ethers.providers.TransactionRequest,
-  ): Promise<string> {
-    const result = await this.viemPublicClient.call({
-      to: transaction.to as `0x${string}`,
-      data: transaction.data as `0x${string}`,
-      value: transaction.value
-        ? BigInt(transaction.value.toString())
-        : undefined,
-    });
-    // viem may return "0x..." (string) or {data: "0x..."} (object)
-    if (typeof result === 'object' && result && 'data' in result) {
-      return result.data as string;
-    }
-    return result as string;
-  }
-
-  async estimateGas(
-    transaction: ethers.providers.TransactionRequest,
-  ): Promise<ethers.BigNumber> {
-    const viemTransaction = {
-      to: transaction.to as `0x${string}`,
-      data: transaction.data as `0x${string}`,
-      value: transaction.value
-        ? BigInt(transaction.value.toString())
-        : undefined,
-    };
-    const gas = await this.viemPublicClient.estimateGas(viemTransaction);
-    return ethers.BigNumber.from(gas.toString());
-  }
-
-  async getGasPrice(): Promise<ethers.BigNumber> {
-    const gasPrice = await this.viemPublicClient.getGasPrice();
-    return ethers.BigNumber.from(gasPrice.toString());
-  }
-
-  async getBlock(blockHashOrBlockTag: string | number): Promise<ViemBlock> {
-    if (
-      typeof blockHashOrBlockTag === 'string' &&
-      blockHashOrBlockTag.startsWith('0x')
-    ) {
-      return await this.viemPublicClient.getBlock({
-        blockHash: blockHashOrBlockTag as `0x${string}`,
-      });
-    } else {
-      return await this.viemPublicClient.getBlock({
-        blockNumber: BigInt(blockHashOrBlockTag),
-      });
-    }
-  }
-
-  async getFeeData(): Promise<ethers.providers.FeeData> {
-    const feeData = await this.viemPublicClient.getFeeHistory({
-      blockCount: 4,
-      blockTag: 'latest' as const,
-      rewardPercentiles: [25, 50, 75],
-    });
-    // Use the latest base fee and priority fee
-    const latestBaseFee =
-      feeData.baseFeePerGas[feeData.baseFeePerGas.length - 1];
-    const latestReward = feeData.reward?.[feeData.reward.length - 1];
-    const medianPriorityFee = latestReward ? latestReward[1] : BigInt(0);
-
-    return {
-      maxFeePerGas: ethers.BigNumber.from(
-        (latestBaseFee + medianPriorityFee).toString(),
-      ),
-      maxPriorityFeePerGas: ethers.BigNumber.from(medianPriorityFee.toString()),
-      gasPrice: ethers.BigNumber.from(latestBaseFee.toString()),
-      lastBaseFeePerGas: ethers.BigNumber.from(latestBaseFee.toString()),
-    };
-  }
-
-  async resolveName(name: string): Promise<string | null> {
-    try {
-      return await this.viemPublicClient.getEnsAddress({
-        name: name as string,
-      });
-    } catch {
-      return null;
-    }
-  }
-
-  async lookupAddress(address: string): Promise<string | null> {
-    try {
-      return await this.viemPublicClient.getEnsName({
-        address: address as `0x${string}`,
-      });
-    } catch {
-      return null;
-    }
-  }
-}
-
-/**
- * Base class for signers that wrap viem Account objects
- *
- * This class contains all the shared implementation logic between taco and taco-auth packages.
- * Package-specific signer classes can extend this to add additional methods.
- */
-export abstract class ViemSignerBase {
-  protected viemAccount: Account;
-  public readonly provider?: ViemProviderBase | undefined;
-
-  // Ethers.js compatibility properties for contract validation
-  readonly _isSigner: boolean = true;
-
-  constructor(viemAccount: Account, provider?: ViemProviderBase | undefined) {
-    this.viemAccount = viemAccount;
-    this.provider = provider;
-  }
-
-  async getAddress(): Promise<string> {
-    return this.viemAccount.address;
-  }
-
-  async signMessage(message: string | Uint8Array): Promise<string> {
-    await checkViemAvailability();
-    if (!this.viemAccount.signMessage) {
-      throw new Error('Account does not support message signing');
-    }
-    // Convert message to hex if it's Uint8Array for compatibility
-    const messageToSign =
-      typeof message === 'string' ? message : ethers.utils.hexlify(message);
-    return await this.viemAccount.signMessage({ message: messageToSign });
-  }
-
-  async getBalance(): Promise<ethers.BigNumber> {
-    if (!this.provider) {
-      throw new Error('Provider is required for getBalance operation');
-    }
-    return await this.provider.getBalance(this.viemAccount.address);
-  }
-
-  async getTransactionCount(): Promise<number> {
-    if (!this.provider) {
-      throw new Error('Provider is required for getTransactionCount operation');
-    }
-    return await this.provider.getTransactionCount(this.viemAccount.address);
-  }
-
-  async estimateGas(
-    transaction: ethers.providers.TransactionRequest,
-  ): Promise<ethers.BigNumber> {
-    if (!this.provider) {
-      throw new Error('Provider is required for estimateGas operation');
-    }
-    return await this.provider.estimateGas(transaction);
-  }
-
-  async call(
-    transaction: ethers.providers.TransactionRequest,
-  ): Promise<string> {
-    if (!this.provider) {
-      throw new Error('Provider is required for call operation');
-    }
-    return await this.provider.call(transaction);
-  }
-
-  async signTypedData(
-    domain: ViemTypedDataDomain,
-    types: Record<string, readonly ViemTypedDataParameter[]>,
-    message: Record<string, unknown>,
-  ): Promise<string> {
-    await checkViemAvailability();
-    if (!this.viemAccount.signTypedData) {
-      throw new Error('Account does not support typed data signing');
-    }
-
-    return await this.viemAccount.signTypedData({
-      domain,
-      types,
-      message,
-      primaryType:
-        Object.keys(types).find((key) => key !== 'EIP712Domain') || 'Message',
-    });
-  }
-
-  /**
-   * Connect this signer to a different provider
-   * This method should be implemented by concrete signer classes
-   */
-  abstract connect(provider?: ViemProviderBase): ViemSignerBase;
 }
